In signer plugin, rename `signer` pkg to `backends`
Update import paths in python code and config files accordingly.
Prevents import paths like
`kinto_remote_settings.signer.signer.<whatever>`.
