+++ config/example.ini
@@ -157,7 +157,7 @@ kinto.signer.reviewers_group = {collection_id}-reviewers
 
 kinto.signer.group_check_enabled = true
 kinto.signer.to_review_enabled = true
-kinto.signer.signer_backend = kinto_remote_settings.signer.signer.autograph
+kinto.signer.signer_backend = kinto_remote_settings.signer.backends.autograph
 
 # Autograph credentials
 # https://github.com/mozilla-services/autograph/blob/29a206fb6/autograph.yaml#L1218-L1221
