+++ kinto_remote_settings/signer/tests/config/signer.ini
@@ -21,7 +21,7 @@ kinto.signer.resources =
     /buckets/bob/collections/source-> /buckets/bob/collections/destination
     /buckets/stage -> /buckets/preview -> /buckets/prod
 
-kinto.signer.signer_backend = kinto_remote_settings.signer.signer.autograph
+kinto.signer.signer_backend = kinto_remote_settings.signer.backends.autograph
 kinto.signer.autograph.server_url = http://localhost:8000
 kinto.signer.autograph.hawk_id = alice
 kinto.signer.autograph.hawk_secret = fs5wgcer9qj819kfptdlp8gm227ewxnzvsuj9ztycsx08hfhzu
